+++ turn-server/Cargo.toml
@@ -23,7 +23,6 @@ categories = [
 [dependencies]
 bytes = "1.4.0"
 rustls = "0.21.2"
-faster-stun = { path = "../stun", version = "1" }
 turn-rs = { path = "../turn", version = "1" }
 clap = { version = "4", features = ["derive", "env"] }
 tokio = { version = "1", features = ["full"] }
